+++ src/main/java/com/wakeUpTogetUp/togetUp/api/alarm/AlarmService.java
@@ -5,8 +5,8 @@ import com.wakeUpTogetUp.togetUp.api.alarm.dto.request.PostAlarmReq;
 import com.wakeUpTogetUp.togetUp.api.alarm.dto.response.AlarmSimpleRes;
 import com.wakeUpTogetUp.togetUp.api.alarm.dto.response.AlarmTimeLineRes;
 import com.wakeUpTogetUp.togetUp.api.alarm.model.Alarm;
-import com.wakeUpTogetUp.togetUp.api.mission.MissionObjectRepository;
-import com.wakeUpTogetUp.togetUp.api.mission.MissionRepository;
+import com.wakeUpTogetUp.togetUp.api.mission.repository.MissionObjectRepository;
+import com.wakeUpTogetUp.togetUp.api.mission.repository.MissionRepository;
 import com.wakeUpTogetUp.togetUp.api.mission.model.Mission;
 import com.wakeUpTogetUp.togetUp.api.mission.model.MissionObject;
 import com.wakeUpTogetUp.togetUp.api.users.UserRepository;
